refactor: Replace NewRedisBinary with redis_executable.NewRedisExecutable. Update usage in all test files.
