+++ internal/test_get_set.go
@@ -2,6 +2,7 @@ package internal
 
 import (
 	"github.com/codecrafters-io/redis-tester/internal/instrumented_resp_client"
+	"github.com/codecrafters-io/redis-tester/internal/redis_executable"
 	"github.com/codecrafters-io/redis-tester/internal/resp_assertions"
 	"github.com/codecrafters-io/redis-tester/internal/test_cases"
 	"github.com/codecrafters-io/tester-utils/random"
@@ -10,7 +11,7 @@ import (
 
 // Tests 'GET, SET'
 func testGetSet(stageHarness *test_case_harness.TestCaseHarness) error {
-	b := NewRedisBinary(stageHarness)
+	b := redis_executable.NewRedisExecutable(stageHarness)
 	if err := b.Run(); err != nil {
 		return err
 	}
