+++ internal/test_ping_pong.go
@@ -2,6 +2,7 @@ package internal
 
 import (
 	"github.com/codecrafters-io/redis-tester/internal/instrumented_resp_client"
+	"github.com/codecrafters-io/redis-tester/internal/redis_executable"
 	resp_client "github.com/codecrafters-io/redis-tester/internal/resp/client"
 	"github.com/codecrafters-io/redis-tester/internal/resp_assertions"
 	"github.com/codecrafters-io/redis-tester/internal/test_cases"
@@ -10,7 +11,7 @@ import (
 )
 
 func testPingPongOnce(stageHarness *test_case_harness.TestCaseHarness) error {
-	b := NewRedisBinary(stageHarness)
+	b := redis_executable.NewRedisExecutable(stageHarness)
 	if err := b.Run(); err != nil {
 		return err
 	}
@@ -39,7 +40,7 @@ func testPingPongOnce(stageHarness *test_case_harness.TestCaseHarness) error {
 }
 
 func testPingPongMultiple(stageHarness *test_case_harness.TestCaseHarness) error {
-	b := NewRedisBinary(stageHarness)
+	b := redis_executable.NewRedisExecutable(stageHarness)
 	if err := b.Run(); err != nil {
 		return err
 	}
@@ -63,7 +64,7 @@ func testPingPongMultiple(stageHarness *test_case_harness.TestCaseHarness) error
 }
 
 func testPingPongConcurrent(stageHarness *test_case_harness.TestCaseHarness) error {
-	b := NewRedisBinary(stageHarness)
+	b := redis_executable.NewRedisExecutable(stageHarness)
 	if err := b.Run(); err != nil {
 		return err
 	}
