+++ internal/test_repl_id.go
@@ -2,6 +2,7 @@ package internal
 
 import (
 	"fmt"
+	"github.com/codecrafters-io/redis-tester/internal/redis_executable"
 	"strings"
 
 	"github.com/codecrafters-io/tester-utils/test_case_harness"
@@ -9,7 +10,7 @@ import (
 
 func testReplReplicationID(stageHarness *test_case_harness.TestCaseHarness) error {
 	deleteRDBfile()
-	b := NewRedisBinary(stageHarness)
+	b := redis_executable.NewRedisExecutable(stageHarness)
 
 	if err := b.Run(); err != nil {
 		return err
