+++ internal/test_repl_info.go
@@ -2,13 +2,14 @@ package internal
 
 import (
 	"fmt"
+	"github.com/codecrafters-io/redis-tester/internal/redis_executable"
 	"strings"
 
 	"github.com/codecrafters-io/tester-utils/test_case_harness"
 )
 
 func testReplInfo(stageHarness *test_case_harness.TestCaseHarness) error {
-	b := NewRedisBinary(stageHarness)
+	b := redis_executable.NewRedisExecutable(stageHarness)
 
 	if err := b.Run(); err != nil {
 		return err
