+++ internal/test_repl_master_cmd_prop.go
@@ -1,6 +1,7 @@
 package internal
 
 import (
+	"github.com/codecrafters-io/redis-tester/internal/redis_executable"
 	"strings"
 
 	"github.com/codecrafters-io/redis-tester/internal/instrumented_resp_client"
@@ -14,12 +15,10 @@ func testReplMasterCmdProp(stageHarness *test_case_harness.TestCaseHarness) erro
 	deleteRDBfile()
 
 	// Run the user's code as a master
-	masterBinary := NewRedisBinary(stageHarness)
-	masterBinary.args = []string{
+	masterBinary := redis_executable.NewRedisExecutable(stageHarness)
+	if err := masterBinary.Run([]string{
 		"--port", "6379",
-	}
-
-	if err := masterBinary.Run(); err != nil {
+	}...); err != nil {
 		return err
 	}
