+++ internal/test_repl_master_psync.go
@@ -2,17 +2,16 @@ package internal
 
 import (
 	"fmt"
+	"github.com/codecrafters-io/redis-tester/internal/redis_executable"
 
 	"github.com/codecrafters-io/tester-utils/test_case_harness"
 )
 
 func testReplMasterPsync(stageHarness *test_case_harness.TestCaseHarness) error {
-	master := NewRedisBinary(stageHarness)
-	master.args = []string{
+	master := redis_executable.NewRedisExecutable(stageHarness)
+	if err := master.Run([]string{
 		"--port", "6379",
-	}
-
-	if err := master.Run(); err != nil {
+	}...); err != nil {
 		return err
 	}
