+++ internal/test_repl_multiple_replicas.go
@@ -2,18 +2,17 @@ package internal
 
 import (
 	"fmt"
+	"github.com/codecrafters-io/redis-tester/internal/redis_executable"
 
 	"github.com/codecrafters-io/tester-utils/test_case_harness"
 )
 
 func testReplMultipleReplicas(stageHarness *test_case_harness.TestCaseHarness) error {
 	deleteRDBfile()
-	master := NewRedisBinary(stageHarness)
-	master.args = []string{
+	master := redis_executable.NewRedisExecutable(stageHarness)
+	if err := master.Run([]string{
 		"--port", "6379",
-	}
-
-	if err := master.Run(); err != nil {
+	}...); err != nil {
 		return err
 	}
