+++ internal/test_repl_wait.go
@@ -2,6 +2,7 @@ package internal
 
 import (
 	"fmt"
+	"github.com/codecrafters-io/redis-tester/internal/redis_executable"
 	"math"
 	"strconv"
 	"time"
@@ -41,12 +42,10 @@ func testWait(stageHarness *test_case_harness.TestCaseHarness) error {
 	deleteRDBfile()
 
 	// Step 1: Boot the user's code as a Redis master.
-	master := NewRedisBinary(stageHarness)
-	master.args = []string{
+	master := redis_executable.NewRedisExecutable(stageHarness)
+	if err := master.Run([]string{
 		"--port", "6379",
-	}
-
-	if err := master.Run(); err != nil {
+	}...); err != nil {
 		return err
 	}
