+++ internal/test_streams_xadd.go
@@ -2,6 +2,7 @@ package internal
 
 import (
 	"fmt"
+	"github.com/codecrafters-io/redis-tester/internal/redis_executable"
 	"strings"
 
 	"github.com/codecrafters-io/tester-utils/logger"
@@ -61,7 +62,7 @@ func (t *XADDTest) Run(client *redis.Client, logger *logger.Logger) error {
 }
 
 func testStreamsXadd(stageHarness *test_case_harness.TestCaseHarness) error {
-	b := NewRedisBinary(stageHarness)
+	b := redis_executable.NewRedisExecutable(stageHarness)
 	if err := b.Run(); err != nil {
 		return err
 	}
