+++ internal/test_streams_xadd_partial_autoid.go
@@ -1,12 +1,13 @@
 package internal
 
 import (
+	"github.com/codecrafters-io/redis-tester/internal/redis_executable"
 	testerutils_random "github.com/codecrafters-io/tester-utils/random"
 	"github.com/codecrafters-io/tester-utils/test_case_harness"
 )
 
 func testStreamsXaddPartialAutoid(stageHarness *test_case_harness.TestCaseHarness) error {
-	b := NewRedisBinary(stageHarness)
+	b := redis_executable.NewRedisExecutable(stageHarness)
 	if err := b.Run(); err != nil {
 		return err
 	}
