+++ internal/test_streams_xread_multiple.go
@@ -1,6 +1,7 @@
 package internal
 
 import (
+	"github.com/codecrafters-io/redis-tester/internal/redis_executable"
 	"strconv"
 
 	testerutils_random "github.com/codecrafters-io/tester-utils/random"
@@ -9,7 +10,7 @@ import (
 )
 
 func testStreamsXreadMultiple(stageHarness *test_case_harness.TestCaseHarness) error {
-	b := NewRedisBinary(stageHarness)
+	b := redis_executable.NewRedisExecutable(stageHarness)
 	if err := b.Run(); err != nil {
 		return err
 	}
